Initial reorganizaiton of assignment 02 ProcessSimulator for GitHub classrooms
Initial reorganization of the assignment 02.  We have reorganized the build,
put in the directory structure, and added in GitHub workflow and vscode
dev container information.

We have begun updates of unit tests for this assignment.  They could use
some more work, and the assignment description will need to be
reviewed/edited for changes.