Fix test for running on GitHub workflow
Because of how tests are run individually in the GitHub
autograder/workflow, we have to make sure that all of the
simulation state is redone at start of each test.

This solution I believe is working.  But all of this really needs to be
converted intot fixtures of some kind, instead of copy-pasta
all of the state at beginning of each test case.