Make task 1 and `isInState()` clearer and a bit easier
Improve task 1 clarity a bit.  Make it explicit that the
`cpu` member variable needs to be initialized, and test it.
Also as we have been meaning to do, give a meaningful
error message for the `isInState()` so that if it does fail,
the student will get more information about what value is
not correct of the expected state that fails a test.